+++ Assets/GeneralScripts/IDamageable.cs
@@ -2,7 +2,7 @@ namespace GeneralScripts
 {
     public interface IDamageable
     {
-        public void TakeDamage(int damage);
+        public void TakeDamage(float damage);
         
         public void OnDeath();
     }
